Move jtag common code into common
